Check prior to down-casting the stream's read length
In this commit, all length sizes prior to m_stream's
read() are of 'size_t' types, however read() only takes in
a signed int.

So this asserts that the incoming read length is small
enough to fit within the signed int when passed into
m_stream's read() call.

assert should only be used for programmer-side logic (as
opposed to user-side data), and in this case the size of
the buffer is programmer controlled (and yes: extremely
unlikely anyone will use a buffer greater than 2 GiB, but
if they do - then atleast this assert will alert them
instead of the read-length overflowing to a new number).